intValue exit LabelDelegate.setAllTracker NAN als NumberCodec type
git-svn-id: https://svn.science.uu.nl/repos/project.fisme.java/Euclides/euclides-parent/trunk@19538 58dd2db4-0f11-0410-9df4-da15f00f50bd

+++ euclides-core/src/main/java/fi/euclides/model/Label.java
@@ -141,7 +141,7 @@ public class Label extends Destroyable implements Observer {
 		// state/value
 		Numbers n;
 		n = codec.readNumber();
-		state = (int)n.longValue();
+		state = n.intValue();
 		setValue( codec.readNumber() );
 	}
